changes size_t to int to match MPI expectations and fixes linking and fixes test launch to use mpiexec

+++ tests/utils/array.hh
@@ -38,7 +38,7 @@ class Data_ {
   // Constructors are all protected -- do not use this class directly!
   //
   // construct from a total length
-  Data_(size_t len)
+  Data_(int len)
       : len_(len),
         do_(std::shared_ptr<T>(new T[len], std::default_delete<T[]>()))
   {
@@ -46,14 +46,14 @@ class Data_ {
   }
 
   // construct and initialize
-  Data_(size_t len, T d)
+  Data_(int len, T d)
       : Data_(len)
   {
     *this = d;
   }
 
   // construct non-owning view
-  Data_(size_t len, T* d)
+  Data_(int len, T* d)
       : len_(len),
         d_(d) {}
 
@@ -62,7 +62,7 @@ class Data_ {
 
  protected:
   // global length
-  const size_t len_;
+  const int len_;
 
   // owning data
   std::shared_ptr<T> do_;
@@ -83,15 +83,15 @@ class Array<T,1> : public Data_<T> {
 
  public:
   // forward construction
-  Array(size_t N)
+  Array(int N)
       : Data_<T>(N) {}
 
   // forward construction
-  Array(size_t N, T t)
+  Array(int N, T t)
       : Data_<T>(N, t) {}
   
   // forward construction
-  Array(size_t N, T* d) :
+  Array(int N, T* d) :
       Data_<T>(N, d)
   {}
 
@@ -99,19 +99,19 @@ class Array<T,1> : public Data_<T> {
   Array(const Array<T,1>& other) = default;
 
   // accessors
-  T& operator()(const size_t i) { assert(0 <= i && i < len_); return d_[i]; }
-  const T& operator()(const size_t i) const { assert(0 <= i && i < len_); return d_[i]; }
+  T& operator()(const int i) { assert(0 <= i && i < len_); return d_[i]; }
+  const T& operator()(const int i) const { assert(0 <= i && i < len_); return d_[i]; }
 
-  T& operator[](const size_t i) { assert(0 <= i && i < len_); return d_[i]; }
-  const T& operator[](size_t i) const { assert(0 <= i && i < len_); return d_[i]; }
+  T& operator[](const int i) { assert(0 <= i && i < len_); return d_[i]; }
+  const T& operator[](int i) const { assert(0 <= i && i < len_); return d_[i]; }
 
   // shape
-  size_t extent(size_t d) const {
+  int extent(int d) const {
     assert(d < 1 && "Array::extent requested for dimension greater than is stored.");
     return shape()[d];
   }
   
-  std::array<size_t,1> shape() const {
+  std::array<int,1> shape() const {
     return { len_ };
   }
   
@@ -128,21 +128,21 @@ class Array<T,2> : public Data_<T> {
 
  public:
   // forward construction
-  Array(size_t M, size_t N)
+  Array(int M, int N)
       : Data_<T>(N*M),
         M_(M),
         N_(N)
   {}
 
   // forward construction
-  Array(size_t M, size_t N, T t)
+  Array(int M, int N, T t)
       : Data_<T>(N*M, t),
         M_(M),
         N_(N)
   {}
   
   // forward construction
-  Array(size_t M, size_t N, T* d)
+  Array(int M, int N, T* d)
       : Data_<T>(N*M, d),
         M_(M),
         N_(N)
@@ -151,25 +151,25 @@ class Array<T,2> : public Data_<T> {
   // forward construction
   Array(const Array<T,2>& other) = default;
 
-  T& operator()(size_t i, size_t j) { assert(0 <= i && i < M_ && 0 <= j && j < N_); return d_[j+i*N_]; }
-  const T& operator()(size_t i, size_t j) const { assert(0 <= i && i < M_ && 0 <= j && j < N_); return d_[j+i*N_]; }
+  T& operator()(int i, int j) { assert(0 <= i && i < M_ && 0 <= j && j < N_); return d_[j+i*N_]; }
+  const T& operator()(int i, int j) const { assert(0 <= i && i < M_ && 0 <= j && j < N_); return d_[j+i*N_]; }
 
-  Array<T,1> operator[](size_t i) { assert(0 <= i && i < M_); return Array<T,1>(N_, &d_[i*N_]); }
-  const Array<T,1> operator[](size_t i) const { assert(0 <= i && i < M_); return Array<T,1>(N_, &d_[i*N_]); }
+  Array<T,1> operator[](int i) { assert(0 <= i && i < M_); return Array<T,1>(N_, &d_[i*N_]); }
+  const Array<T,1> operator[](int i) const { assert(0 <= i && i < M_); return Array<T,1>(N_, &d_[i*N_]); }
 
   // shape
-  size_t extent(size_t d) const {
+  int extent(size_t d) const {
     assert(d < 2 && "Array::extent requested for dimension greater than is stored.");
     return shape()[d];
   }
   
-  std::array<size_t,2> shape() const {
+  std::array<int,2> shape() const {
     return { M_, N_ };
   }
 
   
  protected:
-  size_t M_, N_;
+  int M_, N_;
   using Data_<T>::d_;
 
 };
@@ -182,7 +182,7 @@ class Array<T,3> : public Data_<T> {
   
  public:
   // forward construction
-  Array(size_t M, size_t N, size_t P)
+  Array(int M, int N, int P)
       : Data_<T>(N*M*P),
         M_(M),
         N_(N),
@@ -190,7 +190,7 @@ class Array<T,3> : public Data_<T> {
   {}
 
   // forward construction
-  Array(size_t M, size_t N, size_t P, T t)
+  Array(int M, int N, int P, T t)
       : Data_<T>(N*M*P, t),
         M_(M),
         N_(N),
@@ -198,7 +198,7 @@ class Array<T,3> : public Data_<T> {
   {}
   
   // forward construction
-  Array(size_t M, size_t N, size_t P, T* d)
+  Array(int M, int N, int P, T* d)
       : Data_<T>(N*M*P, d),
         M_(M),
         N_(N),
@@ -208,40 +208,40 @@ class Array<T,3> : public Data_<T> {
   // forward construction
   Array(const Array<T,3>& other) = default;
 
-  T& operator()(size_t i, size_t j, size_t k) {
+  T& operator()(int i, int j, int k) {
     assert(0 <= i && i < M_ &&
            0 <= j && j < N_ &&
            0 <= k && k < P_);
     return d_[k + P_*(j+i*N_)];
   }
-  const T& operator()(size_t i, size_t j, size_t k) const {
+  const T& operator()(int i, int j, int k) const {
     assert(0 <= i && i < M_ &&
            0 <= j && j < N_ &&
            0 <= k && k < P_);
     return d_[k + P_*(j+i*N_)];
   }
 
-  Array<T,2> operator[](size_t i) {
+  Array<T,2> operator[](int i) {
     assert(0 <= i && i < M_);
     return Array<T,2>(N_, P_, &d_[i*N_*P_]);
   }
-  const Array<T,2> operator[](size_t i) const {
+  const Array<T,2> operator[](int i) const {
     assert(0 <= i && i < M_);
     return Array<T,2>(N_, P_, &d_[i*N_*P_]);
   }
 
   // shape
-  size_t extent(size_t d) const {
+  int extent(int d) const {
     assert(d < 3 && "Array::extent requested for dimension greater than is stored.");
     return shape()[d];
   }
   
-  std::array<size_t,3> shape() const {
+  std::array<int,3> shape() const {
     return { M_, N_, P_ };
   }
   
  protected:
-  size_t M_, N_, P_;
+  int M_, N_, P_;
   using Data_<T>::d_;
 
 };
@@ -254,7 +254,7 @@ class Array<T,4> : public Data_<T> {
   static const size_t dim = 4;
  public:
   // forward construction
-  Array(size_t M, size_t N, size_t P, size_t Q)
+  Array(int M, int N, int P, int Q)
       : Data_<T>(N*M*P*Q),
         M_(M),
         N_(N),
@@ -263,7 +263,7 @@ class Array<T,4> : public Data_<T> {
   {}
 
   // forward construction
-  Array(size_t M, size_t N, size_t P, size_t Q, T t)
+  Array(int M, int N, int P, int Q, T t)
       : Data_<T>(N*M*P*Q, t),
         M_(M),
         N_(N),
@@ -272,7 +272,7 @@ class Array<T,4> : public Data_<T> {
   {}
   
   // forward construction
-  Array(size_t M, size_t N, size_t P, size_t Q, T* d)
+  Array(int M, int N, int P, int Q, T* d)
       : Data_<T>(N*M*P*Q, d),
         M_(M),
         N_(N),
@@ -283,14 +283,14 @@ class Array<T,4> : public Data_<T> {
   // forward construction
   Array(const Array<T,4>& other) = default;
 
-  T& operator()(size_t i, size_t j, size_t k, size_t l) {
+  T& operator()(int i, int j, int k, int l) {
     assert(0 <= i && i < M_ &&
            0 <= j && j < N_ &&
            0 <= k && k < P_ &&
            0 <= l && l < Q_);
     return d_[l + Q_*(k + P_*(j+i*N_))];
   }
-  const T& operator()(size_t i, size_t j, size_t k, size_t l) const {
+  const T& operator()(int i, int j, int k, int l) const {
     assert(0 <= i && i < M_ &&
            0 <= j && j < N_ &&
            0 <= k && k < P_ &&
@@ -298,27 +298,27 @@ class Array<T,4> : public Data_<T> {
     return d_[l + Q_*(k + P_*(j+i*N_))];
   }
 
-  Array<T,3> operator[](size_t i) {
+  Array<T,3> operator[](int i) {
     assert(0 <= i && i < M_);
     return Array<T,3>(N_, P_, Q_, &d_[i*N_*P_*Q_]);
   }
-  const Array<T,3> operator[](size_t i) const {
+  const Array<T,3> operator[](int i) const {
     assert(0 <= i && i < M_);
     return Array<T,3>(N_, P_, Q_, &d_[i*N_*P_*Q_]);
   }
 
   // shape
-  size_t extent(size_t d) const {
+  int extent(int d) const {
     assert(d < 4 && "Array::extent requested for dimension greater than is stored.");
     return shape()[d];
   }
   
-  std::array<size_t,4> shape() const {
+  std::array<int,4> shape() const {
     return { M_, N_, P_, Q_ };
   }
   
  protected:
-  size_t M_, N_, P_, Q_;
+  int M_, N_, P_, Q_;
   using Data_<T>::d_;
 
 };
@@ -331,8 +331,8 @@ class Array<T,4> : public Data_<T> {
 // NOTE, this does no transposing!
 //
 template<size_t D1, size_t D2, typename T>
-Array<T,D2> reshape(Array<T,D1>& arr_in, const std::array<size_t, D2>& new_shape) {
-  size_t new_length = std::accumulate(new_shape.begin(), new_shape.end(), 1, std::multiplies<size_t>());
+Array<T,D2> reshape(Array<T,D1>& arr_in, const std::array<int, D2>& new_shape) {
+  int new_length = std::accumulate(new_shape.begin(), new_shape.end(), 1, std::multiplies<int>());
   if (new_length != arr_in.size()) {
     //std::stringstream err;
     std::cout << "Invalid Array reshape, cannot reshape object of size: " << arr_in.size() << " into array of size: " << new_length;
@@ -355,7 +355,7 @@ void
 deep_copy(Array_type& arr, const Array<T,1>& arr_in)
 {
   assert(arr.extent(0) == arr_in.extent(0));
-  for (size_t i=0; i!=arr_in.extent(0); ++i) {
+  for (int i=0; i!=arr_in.extent(0); ++i) {
     arr(i) = arr_in(i);
   }
 }
@@ -366,8 +366,8 @@ deep_copy(Array_type& arr, const Array<T,2>& arr_in)
 {
   assert(arr.extent(0) == arr_in.extent(0));
   assert(arr.extent(1) == arr_in.extent(1));
-  for (size_t i=0; i!=arr_in.extent(0); ++i) {
-    for (size_t j=0; j!=arr_in.extent(1); ++j) {
+  for (int i=0; i!=arr_in.extent(0); ++i) {
+    for (int j=0; j!=arr_in.extent(1); ++j) {
       arr(i,j) = arr_in(i,j);
     }
   }
@@ -380,9 +380,9 @@ deep_copy(Array_type& arr, const Array<T,3>& arr_in)
   assert(arr.extent(0) == arr_in.extent(0));
   assert(arr.extent(1) == arr_in.extent(1));
   assert(arr.extent(2) == arr_in.extent(2));
-  for (size_t i=0; i!=arr_in.extent(0); ++i) {
-    for (size_t j=0; j!=arr_in.extent(1); ++j) {
-      for (size_t k=0; k!=arr_in.extent(2); ++k) {
+  for (int i=0; i!=arr_in.extent(0); ++i) {
+    for (int j=0; j!=arr_in.extent(1); ++j) {
+      for (int k=0; k!=arr_in.extent(2); ++k) {
         arr(i,j,k) = arr_in(i,j,k);
       }
     }
@@ -397,10 +397,10 @@ deep_copy(Array_type& arr, const Array<T,4>& arr_in)
   assert(arr.extent(1) == arr_in.extent(1));
   assert(arr.extent(2) == arr_in.extent(2));
   assert(arr.extent(3) == arr_in.extent(3));
-  for (size_t i=0; i!=arr_in.extent(0); ++i) {
-    for (size_t j=0; j!=arr_in.extent(1); ++j) {
-      for (size_t k=0; k!=arr_in.extent(2); ++k) {
-        for (size_t l=0; l!=arr_in.extent(3); ++l) {
+  for (int i=0; i!=arr_in.extent(0); ++i) {
+    for (int j=0; j!=arr_in.extent(1); ++j) {
+      for (int k=0; k!=arr_in.extent(2); ++k) {
+        for (int l=0; l!=arr_in.extent(3); ++l) {
           arr(i,j,k,l) = arr_in(i,j,k,l);
         }
       }
